chore: Update CI/CD workflow and versioned library imports in base template
* Update CI workflows:
  - Rename `docker` job to `build_latest`
  - Add new `build_hash` job with same steps as `build_latest`
  - Add new `deploy` job to pull latest image and recreate containers
  - Update branch list for CI triggering
  - Add new step in `build_latest` and `build_hash` jobs to set `SHORT_SHA` env variable
  * Update base template:
  - Update import statement for `libs` in `gargantua/templates/base.html`
  - Replace import statement with versioned import from `https://s3.laisky.com/static/laisky/libs.js?v=20240326-v3`
  - Apply versioned import for `libs` and update `LoadJsModules` call for site-specific JS modules

This commit updates the CI workflows by renaming the `docker` job, adding new jobs for building and deploying, and updating the branch list for CI triggering. It also sets a new `SHORT_SHA` environment variable. The base template is updated with a versioned import for `libs` and updates to the `LoadJsModules` call for site-specific JS modules.